Zwift Team Results Slack "Slash Command"
This was an interesting problem to solve. Using the largely undocument ZwiftPower API (thanks folks!) I am able to pull the results for a given event and show the results for just our team.

The `SlackZwitResultsTrigger` function is invoked by a slack slash command that I manually set up for our workspace. Because the ZwiftPower API can be very slow and due to the fact that Slack only gives you 3 seconds to respond to an initial command, this entry point simply responds to the user and pushes the message into a azure queue, where it is picked up at a later point by the `ResultsWorker`.

The `ResultsWorker` receives a message, grabs the results from the zp API, and responds to the user using the data encoded in the queued message.

Obviously this implementation leaves a lot to be desired in terms of configurability, but it works for the short term and can be itterated on.